+++ src/views/examens/planification/components/DetailItemv2.vue
