+++ src/views/examens/planification/components/DetailsItem.vue
